Use ansible to configure remote directories

+++ create_links.sh
@@ -8,7 +8,7 @@ for file in ${file_list}
 do
     if [ -f ${HOME}/${file} ]; then
         cp ${HOME}/${file} ${HOME}/old_dotfiles
-        rm ${file}
+        \rm ${HOME}/${file}
     fi
 done
 for file in ${file_list}
@@ -19,4 +19,8 @@ done
 
 IPY_INIT_DIR=${HOME}/.ipython/profile_default
 mkdir -p ${IPY_INIT_DIR}
-ln -s ${base_dir}/ipython_config.py ${IPY_INIT_DIR}/ipython_config.py
+    if [ -f ${IPY_INIT_DIR}/ipython_config.py ]; then
+        cp  ${IPY_INIT_DIR}/ipython_config.py ${HOME}/old_dotfiles
+        \rm  ${IPY_INIT_DIR}/ipython_config.py
+        ln -s ${base_dir}/ipython_config.py ${IPY_INIT_DIR}/ipython_config.py
+    fi
